metal-surface-defect-detection-cnn-main/requirements.txt
ipykernel==6.29.5
PyYAML==6.0.2
scikit-learn==1.6.1
ensure==1.0.4
streamlit==1.43.2
mlflow==2.20.2
pandas==2.2.3
numpy==1.26.4
matplotlib==3.10.0
seaborn==0.13.2
tqdm==4.67.1
joblib==1.4.2
python-dotenv==1.0.1
statsmodels==0.14.4
python-box==7.3.2
scikeras==0.13.0
kagglehub==0.3.10
tensorflow==2.19.0
tensorboard==2.19.0
keras-tuner==1.4.7
opencv-python==4.11.0.86
#-e .